refactor(utils): Extract each utility function to its own file
