+++ content/utils/getXULElementById.ts
@@ -1,0 +1,5 @@
+export default function getXULElementById<E extends XUL.XULElement>(
+  id: string
+): E {
+  return document.getElementById(id) as unknown as E;
+}
