+++ content/utils/hasErrorStack.ts
@@ -1,0 +1,5 @@
+export default function hasErrorStack(
+  error: unknown
+): error is Required<Error> {
+  return typeof (error as Error).stack !== 'undefined';
+}
